qdb-wal-switch/requirements.txt
psycopg2-binary>=2,<3